requirements.txt
asgiref==3.3.4
certifi==2021.5.30
charset-normalizer==2.0.4
cloudinary==1.26.0
Django==3.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
django-decouple==2.1
djangorestframework==3.12.4
Pillow==8.3.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
